requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
rembg
python-dotenv
gunicorn
